+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The following is a curated list of changes in the Enact project, newest changes 
 ### Changed
 
 - `ui/Scroller` scroll animation method for `scrollMode: native`
+### Fixed
+
+- `spotlight` navigation from an element to a different container on page load.
+
 ## [5.3.1] - 2025-10-14
 
 ### Fixed